Add files via upload
This project is based on Pytorch to study the image classification algorithm and realizes the image classification 
and the recognition of Intel Image Classification by using nn to build a fully connected neural network model. 
As an alternative to Numpy, Pytorch has advanced features and supports GPU acceleration, so it can quickly build 
neural networks and get effective training. The graph structure in Pytorch is easy to understand, and more 
importantly, it is easy to debug for researchers. This article uses the train data set ,the accuracy of this network to 
reach 95.49%, and the image recognition has been able to have a higher accuracy.
❖ Train loss: 0.1732
❖ Train accuracy: 95.49%
❖ Test accuracy: 75.6% max